+++ src/constants/actionTypes.js
@@ -1,3 +1,4 @@
 export const MARKDOWN_CHANGED = 'MARKDOWN_CHANGED'
 export const FILE_LOADED = 'FILE_LOADED'
 export const TOGGLE_EDIT = 'TOGGLE_EDIT'
+export const CHANGE_THEME = 'CHANGE_THEME'